Changed double to float

+++ inc/core/vt_database.h
@@ -23,11 +23,11 @@ void _vt_database_initialize_pearsoncoefficientdb(VT_DATABASE* database_ptr);
 // Store
 uint32_t _vt_database_store_falltime(VT_DATABASE* database_ptr, int fall_time, int sensorid);
 
-uint32_t _vt_database_store_pearsoncoefficient(VT_DATABASE* database_ptr, double pearson_coefficient, int sensorid);
+uint32_t _vt_database_store_pearsoncoefficient(VT_DATABASE* database_ptr, float pearson_coefficient, int sensorid);
 
 // Evaluate
 int _vt_database_evaluate_nrmse(VT_DATABASE* database_ptr, uint32_t* fallcurvearray);
 
-int _vt_database_evaluate_pearson_falltime(VT_DATABASE* database_ptr, int fall_time, double pearson_coefficient);
+int _vt_database_evaluate_pearson_falltime(VT_DATABASE* database_ptr, int fall_time, float pearson_coefficient);
 
 #endif
