+++ inc/core/vt_fingerprint.h
@@ -19,7 +19,7 @@ uint32_t _vt_fingerprint_calculate_falltime_pearsoncoefficient(uint32_t* fingerp
     uint32_t fingerprint_length,
     uint32_t sampling_frequency,
     int* fall_time,
-    double* pearson_coefficient);
+    float* pearson_coefficient);
 
 VT_CURVE_SHAPE _vt_fingerprint_calculate_shape(uint32_t* fingerprint, int fingerprint_length);
 
@@ -28,9 +28,9 @@ uint32_t _vt_fingerprint_calculate_maximum_index(uint32_t* fingerprint, uint32_t
 uint32_t _vt_fingerprint_calculate_37index(uint32_t* fingerprint, uint32_t fingerprint_length);
 
 // Evaluate
-double _vt_fingerprint_evaluate_correlationCoefficient(
+float _vt_fingerprint_evaluate_correlationCoefficient(
     uint32_t* fingerprint1, uint32_t* fingerprint2, int fingeprprint_length);
 
-double _vt_fingerprint_evaluate_nrmse(uint32_t* fingerpint1, uint32_t* fingerprint2, int fingerpint_length);
+float _vt_fingerprint_evaluate_nrmse(uint32_t* fingerpint1, uint32_t* fingerprint2, int fingerpint_length);
 
 #endif
